ported fixes to monitoring fixtures

+++ src/tasks.py
@@ -24,6 +24,7 @@ import json
 import time
 import docker
 import socket
+import ipaddress
 import logging
 import datetime
 
@@ -46,8 +47,8 @@ def waitfordbs(ctx):
 def update(ctx):
     print("***************************setting env*********************************")
     ctx.run("env", pty=True)
-    pub_ip = _geonode_public_host_ip()
-    print(f"Public Hostname or IP is {pub_ip}")
+    pub_host = _geonode_public_host()
+    print(f"Public Hostname or IP is {pub_host}")
     pub_port = _geonode_public_port()
     print(f"Public PORT is {pub_port}")
     pub_protocol = "https" if pub_port == "443" else "http"
@@ -70,18 +71,18 @@ def update(ctx):
         print(f"Can not delete the {override_env} file as it doesn't exists")
 
     if pub_port:
-        siteurl = f"{pub_protocol}://{pub_ip}:{pub_port}/"
-        gs_pub_loc = f"http://{pub_ip}:{pub_port}/geoserver/"
+        siteurl = f"{pub_protocol}://{pub_host}:{pub_port}/"
+        gs_pub_loc = f"http://{pub_host}:{pub_port}/geoserver/"
     else:
-        siteurl = f"{pub_protocol}://{pub_ip}/"
-        gs_pub_loc = f"http://{pub_ip}/geoserver/"
+        siteurl = f"{pub_protocol}://{pub_host}/"
+        gs_pub_loc = f"http://{pub_host}/geoserver/"
     envs = {
         "local_settings": str(_localsettings()),
         "siteurl": os.environ.get("SITEURL", siteurl),
         "geonode_docker_host": geonode_docker_host,
         "public_protocol": pub_protocol,
-        "public_fqdn": str(pub_ip) + str(f":{pub_port}" if pub_port else ""),
-        "public_host": str(pub_ip),
+        "public_fqdn": str(pub_host) + str(f":{pub_port}" if pub_port else ""),
+        "public_host": str(pub_host),
         "dburl": os.environ.get("DATABASE_URL", db_url),
         "geodburl": os.environ.get("GEODATABASE_URL", geodb_url),
         "static_root": os.environ.get("STATIC_ROOT", "/mnt/volumes/statics/static/"),
@@ -125,7 +126,7 @@ def update(ctx):
         )
     except ValueError:
         current_allowed = []
-    current_allowed.extend([str(pub_ip), f"{pub_ip}:{pub_port}"])
+    current_allowed.extend([str(pub_host), f"{pub_host}:{pub_port}"])
     allowed_hosts = [str(c) for c in current_allowed] + ['"geonode"', '"django"']
 
     ctx.run(
@@ -409,17 +410,28 @@ def collectstatic(ctx):
 
 @task
 def monitoringfixture(ctx):
-    print("*******************monitoring fixture********************************")
-    ctx.run("rm -rf /tmp/default_monitoring_apps_docker.json", pty=True)
-    _prepare_monitoring_fixture()
-    try:
-        ctx.run(
-            f"django-admin.py loaddata /tmp/default_monitoring_apps_docker.json \
---settings={_localsettings()}",
-            pty=True,
-        )
-    except Exception as e:
-        logger.error(f"ERROR installing monitoring fixture: {str(e)}")
+    if ast.literal_eval(os.environ.get("MONITORING_ENABLED", "False")):
+        print("*******************monitoring fixture********************************")
+        ctx.run("rm -rf /tmp/default_monitoring_apps_docker.json", pty=True)
+        _prepare_monitoring_fixture()
+        try:
+            ctx.run(
+                f"django-admin loaddata metric_data.json \
+    --settings={_localsettings()}",
+                pty=True,
+            )
+            ctx.run(
+                f"django-admin loaddata notifications.json \
+    --settings={_localsettings()}",
+                pty=True,
+            )
+            ctx.run(
+                f"django-admin loaddata /tmp/default_monitoring_apps_docker.json \
+    --settings={_localsettings()}",
+                pty=True,
+            )
+        except Exception as e:
+            logger.error(f"ERROR installing monitoring fixture: {str(e)}")
 
 
 @task
@@ -486,6 +498,12 @@ address {ip_list[0]}"
         )
     return ip_list[0]
 
+def _is_valid_ip(ip):
+    try:
+        ipaddress.IPv4Address(ip)
+        return True
+    except Exception as e:
+        return False
 
 def _container_exposed_port(component, instname):
     port = "80"
@@ -554,9 +572,16 @@ def _gs_service_availability(url):
         return True
 
 
-def _geonode_public_host_ip():
+def _geonode_public_host():
     gn_pub_hostip = os.getenv("GEONODE_LB_HOST_IP", None)
     if not gn_pub_hostip:
+        gn_pub_hostip = _docker_host_ip()
+    return gn_pub_hostip
+
+
+def _geonode_public_host_ip():
+    gn_pub_hostip = os.getenv("GEONODE_LB_HOST_IP", None)
+    if not gn_pub_hostip or not _is_valid_ip(gn_pub_hostip):
         gn_pub_hostip = _docker_host_ip()
     return gn_pub_hostip
 
@@ -614,7 +639,7 @@ def _prepare_monitoring_fixture():
     # upurl = urlparse(os.environ['SITEURL'])
     # net_scheme = upurl.scheme
     # net_loc = upurl.netloc
-    pub_ip = _geonode_public_host_ip()
+    pub_ip = _geonode_public_host()
     print(f"Public Hostname or IP is {pub_ip}")
     pub_port = _geonode_public_port()
     print(f"Public PORT is {pub_port}")
